move demo to examples/
