+++ silx/gui/widgets/PeriodicTable.py
@@ -699,55 +699,3 @@ class PeriodicList(qt.QTreeWidget):
         """
         for idx in range(len(self.tree_items)):
             self.tree_items[idx].setSelected(_defaultTableItems[idx].symbol in symbolList)
-
-
-def createComboBoxWidget(parent):
-    def change_combo(item):
-        print("New combo selection:", item.symbol)
-
-    widget = qt.QWidget(parent)
-    widget.setLayout(qt.QVBoxLayout())
-    widget.pc = PeriodicCombo(widget)
-    widget.layout().addWidget(widget.pc)
-
-    widget.pc.setSelection("Li")
-    widget.pc.sigSelectionChanged.connect(change_combo)
-
-    return widget
-
-def main():
-    a = qt.QApplication(sys.argv)
-    a.lastWindowClosed.connect(a.quit)
-
-    w = qt.QTabWidget()
-
-    pt = PeriodicTable(w, selectable=True)
-    pc = createComboBoxWidget(w)
-    pl = PeriodicList(w)
-
-    w.addTab(pt, "PeriodicTable")
-    w.addTab(pl, "PeriodicList")
-    w.addTab(pc, "PeriodicCombo")
-
-    pt.setSelection(['H', 'Fe', 'Si'])
-    pl.setSelectedElements(['H', 'Be', 'F'])
-
-    def change_list(items):
-        print("New list selection:", [item.symbol for item in items])
-
-
-    def click_table(item):
-        print("New table click:", item.symbol)
-
-    def change_table(items):
-        print("New table selection:", [item.symbol for item in items])
-
-    pt.sigElementClicked.connect(click_table)
-    pt.sigSelectionChanged.connect(change_table)
-    pl.sigSelectionChanged.connect(change_list)
-
-    w.show()
-    a.exec_()
-
-if __name__ == "__main__":
-    main()
